requirements.txt
flask
psycopg2-binary
argon2-cffi
werkzeug
